cloudwise_finops/requirements.txt
# Core Framework
agency-swarm>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
openai>=1.0.0

# AWS Integration
boto3>=1.34.0
botocore>=1.34.0

# GCP Integration
google-cloud-bigquery>=3.13.0
google-auth>=2.23.0

# Azure Integration
azure-identity>=1.15.0
azure-mgmt-costmanagement>=4.0.0

# Data Processing & Statistical Analysis
pandas>=2.1.0
numpy>=1.24.0
scipy>=1.11.0

# Database (optional PostgreSQL support)
psycopg2-binary>=2.9.0

# Additional dependencies will be added by tools-creator for MCP servers
